lang-graphBasico/requirements.txt
langgraph
langchain_community
langgraph
python-dotenv
ipykernel
langchain-openai